third_party/WebKit/LayoutTests/fast/css/getComputedStyle/counterIncrement-without-counter.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
    <div id="console"></div>
    <script>
        description("Test that the initial value of 'counter-increment' and 'counter-reset' is 'none'.");
        var style = window.getComputedStyle(document.body, null);
        shouldBeEqualToString("style.counterIncrement", "none");
        shouldBeEqualToString("style.counterReset", "none");
    </script>
</body>
</html>
